requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.10.7
gunicorn==20.1.0
huggingface-hub==0.13.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
sympy==1.11.1
tokenizers==0.13.2
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
